Fix comments p.2
